# TheDashboard.FrontendApi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Workshop API",
    "version": "v1"
  },
  "paths": {
    "/api/Dashboard": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "operationId": "GetAll",
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DashboardDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Dashboard"
        ],
        "operationId": "AddDashboard",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DashboardDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DashboardDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DashboardDto"
              }
            }
          }
        },
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Dashboard/{id}": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "operationId": "Get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "maximum": 1999,
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Dashboard"
        ],
        "operationId": "UpdateDashboard",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DashboardDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DashboardDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DashboardDto"
              }
            }
          }
        },
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "404": {
            "description": "Not Found"
          },
          "202": {
            "description": "Accepted"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "delete": {
        "tags": [
          "Dashboard"
        ],
        "operationId": "RemoveDashboard",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "202": {
            "description": "Accepted"
          }
        }
      }
    },
    "/api/Dashboard/search": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "operationId": "Search",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DashboardDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "hasTiles": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      }
    }
  }
}